+++ src/document/italic_line_element.cpp
@@ -1,0 +1,4 @@
+#include "italic_line_element.hpp"
+
+ItalicLineElement::ItalicLineElement(const std::string & content) :
+	LineElement(content) {}